pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GRA_Longevity"
version = "0.1.0"
description = "In silico platform for GRA-based longevity research"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "scipy>=1.7",
    "torch>=1.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "jupyter>=1.0"]
